test(angular): rename utils
