+++ packages/angular/test/base/e2e/utils/drag-utils.ts
